Fixed compiler warnings

+++ dds/DCPS/ConfigStoreImpl.cpp
@@ -527,7 +527,7 @@ ConfigStoreImpl::get(const char* key,
 
   const char* start = t.c_str();
   while (const char* next_comma = std::strchr(start, ',')) {
-    const size_t size = next_comma - start;
+    const size_t size = static_cast<size_t>(next_comma - start);
     retval.push_back(String(start, size));
     start = next_comma + 1;
   }
